getSimplePathsFST
